Version2.0

+++ API/APIREST/Models/Curso.cs
@@ -10,9 +10,9 @@ namespace APIREST.Models
         public int IdCurso { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
-        public decimal Costo { get; set; }
+        public float Costo { get; set; }
         public int? IdInstructor { get; set; }
-        public decimal Duracion { get; set; }
+        public float Duracion { get; set; }
 
         public virtual DatosInstructor IdInstructorNavigation { get; set; }
     }
